feat: add voters/upvoters v3 routes
closes #12423
