+++ public/openapi/write/posts/pid/upvoters.yaml
@@ -1,0 +1,33 @@
+get:
+  tags:
+    - posts
+  summary: get upvoter usernames of a post
+  description: This is used for getting a list of upvoter usernames for the vote tooltip
+  parameters:
+    - in: path
+      name: pid
+      schema:
+        type: string
+      required: true
+      description: a valid post id
+      example: 2
+  responses:
+    '200':
+      description: Usernames of upvoters of post
+      content:
+        application/json:
+          schema:
+            type: object
+            properties:
+              status:
+                $ref: ../../../components/schemas/Status.yaml#/Status
+              response:
+                type: object
+                properties:
+                  otherCount:
+                    type: number
+                  usernames:
+                    type: array
+                  cutoff:
+                    type: number
+
